bloc_2_etl/repository/requirements.txt
pandas
sqlalchemy
psycopg2-binary
python-dotenv
pyarrow
boto3
s3fs
pymongo
botocore
pytest
